% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTSNE.R
\name{digest_args}
\alias{digest_args}
\title{digest_args}
\usage{
digest_args(env = parent.frame(), to_ignore = character(), ...)
}
\arguments{
\item{...}{}
}
\description{
returns digest results of name list of parameters of calling function
}
