% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTSNE.runApp.R
\name{expTSNE.runApp}
\alias{expTSNE.runApp}
\title{expTSNE.runApp}
\usage{
expTSNE.runApp(et)
}
\arguments{
\item{et}{expTSNE object}
}
\value{

}
\description{
expTSNE.runApp
}
\examples{
ex_data = system.file("extdata/test_expTSNE", package = "expTSNE", mustWork = TRUE)
et = expTSNE.load(ex_data)
expTSNE.runApp(et)
}
