% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util_functions.R
\name{get_column_classes}
\alias{get_column_classes}
\title{get class for every column in df}
\usage{
get_column_classes(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
character vector of classes
}
\description{
get class for every column in df
}
