% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util_functions.R
\name{locate_genes_in_df}
\alias{locate_genes_in_df}
\title{locate the first column in df containing valid_genes
enforces gene_name as column name}
\usage{
locate_genes_in_df(df, valid_genes)
}
\arguments{
\item{df}{a data.frame to search}

\item{valid_genes}{character vector of genes to look for}
}
\value{
data.frame with gene names in gene_name var
}
\description{
locate the first column in df containing valid_genes
enforces gene_name as column name
}
