% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTSNE.plot.R
\name{plot_expTSNE}
\alias{plot_expTSNE}
\title{Title}
\usage{
plot_expTSNE(
  et,
  color_var = NULL,
  facet_var = NULL,
  color_palette = "Dark2",
  show_all_points_per_facet = TRUE
)
}
\arguments{
\item{et}{expTSNE object}

\item{color_var}{}

\item{facet_var}{}

\item{color_palette}{}

\item{show_all_points_per_facet}{}
}
\description{
Title
}
\examples{
ex_data = system.file("extdata/test_expTSNE", package = "expTSNE", mustWork = TRUE)
et = expTSNE.load(ex_data)
plot_expTSNE(et)
}
