% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_util_functions.R
\name{validate_genes_in_df}
\alias{validate_genes_in_df}
\title{determine which genes in gene_name are valid}
\usage{
validate_genes_in_df(df, valid_genes, valid_var = "valid")
}
\arguments{
\item{df}{data.frame to inspect and modify}

\item{valid_genes}{character vector of all valid genes}

\item{valid_var}{var name to add validity logical to}
}
\value{
data.frame with additional variable indicating validity (in valid_var) of genes
}
\description{
determine which genes in gene_name are valid
}
