% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTSNE.R
\docType{class}
\name{expTSNE-class}
\alias{expTSNE-class}
\alias{expTSNE}
\title{expTSNE}
\usage{
expTSNE()
}
\value{



}
\description{
expTSNE

expTSNE
}
\section{Slots}{

\describe{
\item{\code{raw_counts}}{matrix.}

\item{\code{norm_counts}}{matrix.}

\item{\code{meta_data}}{data.frame.}

\item{\code{perplexity}}{numeric.}

\item{\code{seed}}{numeric.}

\item{\code{selected_rows}}{character.}

\item{\code{selected_columns}}{character.}
}}

