% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTSNE.R
\name{run_TSNE}
\alias{run_TSNE}
\title{run_TSNE}
\usage{
run_TSNE(counts, apply_normalization = FALSE, perplexity = 30, seed = NULL)
}
\arguments{
\item{counts}{}

\item{perplexity}{}

\item{seed}{}
}
\value{

}
\description{
run_TSNE
}
\examples{
ex_data = system.file("extdata/test_expTSNE.input", package = "expTSNE", mustWork = TRUE)
et = expTSNE.load(ex_data)
tsne_df = run_TSNE(et$norm_counts)
}
